Implement insertion for 2-3 (playing with OCaml)
That wasn't too bad. Strengthened representation to eliminate
'nullability' of nodes which could violate invariants of the tree
structure - distinguish empty trees disjointly from interior tree nodes,
and distinguish 1 and 2 leaves from 2 and 3 interior nodes.

Seems like the hairiness of special cases in order to implement 2-3
trees is overstated, but I'll try a red-black tree to do a side-by-side
comparison. Not sure it makes a difference that I'm doing it
persistently, versus implementing rotations by mutating the tree (yuck).